+++ aggregator/src/light_aggregator.rs
@@ -567,6 +567,7 @@ mod tests {
                 bls12_381: false,
                 base64: false,
                 nr_pow2range_cols: 4,
+                automaton: false,
             }
         }