+++ circuits/examples/bitcoin_signature.rs
@@ -145,6 +145,7 @@ impl Relation for BitcoinSigExample {
             bls12_381: false,
             base64: false,
             nr_pow2range_cols: 4,
+            automaton: false,
         }
     }