+++ circuits/examples/ecc_ops.rs
@@ -77,6 +77,7 @@ impl Relation for EccExample {
             bls12_381: false,
             base64: false,
             nr_pow2range_cols: 1,
+            automaton: false,
         }
     }