+++ circuits/examples/native_gadget.rs
@@ -97,6 +97,7 @@ impl Relation for NativeGadgetExample {
             bls12_381: false,
             base64: false,
             nr_pow2range_cols: 1,
+            automaton: false,
         }
     }