+++ circuits/examples/poseidon.rs
@@ -49,6 +49,7 @@ impl Relation for PoseidonExample {
             bls12_381: false,
             base64: false,
             nr_pow2range_cols: 1,
+            automaton: false,
         }
     }