+++ circuits/examples/sha_preimage.rs
@@ -63,6 +63,7 @@ impl Relation for ShaPreImageCircuit {
             bls12_381: false,
             base64: false,
             nr_pow2range_cols: 1,
+            automaton: false,
         }
     }